refactor rubocop & htmlbeautifier
