+++ app/views/admin/orders/index.html.erb
@@ -7,6 +7,7 @@
       <th>購入者</th>
       <th>メールアドレス</th>
       <th>合計金額</th>
+      <th>割引額</th>
       <th>注文日時</th>
       <th>アクション</th>
     </tr>
