+++ app/views/admin/orders/show.html.erb
@@ -20,7 +20,7 @@
         <td><%= item.name %></td>
         <td><%= number_to_currency(item.price, unit: "$") %></td>
         <td><%= item.quantity %></td>
-        <td><%= number_to_currency(item.price * item.quantity, unit: "¥") %></td>
+        <td><%= number_to_currency(item.price * item.quantity, unit: "$") %></td>
         <td>
           <%# productへの関連がoptionalなのでnilチェックが必要 %>
           <% if item.product %>
@@ -35,7 +35,7 @@
   </tbody>
 </table>
 <% if @order.respond_to?(:discount_amount) && @order.discount_amount.present? && @order.discount_amount > 0 %>
-  <h5>プロモーションコード割引額: <%= number_to_currency(@order.discount_amount, unit: "¥") %></h4>
+  <h5>プロモーションコード割引額: <%= number_to_currency(@order.discount_amount, unit: "$") %></h4>
 <% end %>
-<h3>合計金額: <%= number_to_currency(@order.total_price, unit: "¥") %></h5>
+<h3>合計金額: <%= number_to_currency(@order.total_price, unit: "$") %></h5>
 <%= link_to '一覧に戻る', admin_orders_path %>
