+++ app/views/customer/cart/_cart_item.html.erb
@@ -31,7 +31,6 @@
           <i class="bi bi-arrow-clockwise"></i>
         <% end %>
       <% end %>
-
       <%= button_to remove_item_cart_path(item), 
       method: :delete,
       data: {turbo_confirm: '本当に削除しますか？'},
